Refactored SongShareHomeView and SongSearchView to be their own View components.
