+++ spoti-friends/src/common/views/searchBars/SearchBar.swift
@@ -2,17 +2,16 @@ import SwiftUI
 
 /// A customizable search bar component.
 ///
-/// Displays a text input field with a placeholder, magnifying glass icon on the left,
-/// and a clear button on the right when the user starts typing.
+/// Displays a text input field with a placeholder and a magnifying glass icon on the left.
 ///
 /// - Parameters:
 ///   - placeholderText: The placeholder text displayed in the search bar.
+///   - searchText: The text to search for.
 ///
 /// - Returns: A View that renders the search bar.
 struct SearchBar: View {
     var placeholderText: String
     @Binding var searchText: String
-    @FocusState private var isFocused: Bool
     
     var body: some View {
         HStack {
@@ -21,7 +20,6 @@ struct SearchBar: View {
                       prompt: Text(placeholderText)
                 .foregroundStyle(Color.PresetColour.whiteSecondary)
             )
-            .focused($isFocused)
             .font(.footnote)
             .padding(.vertical, 8)
             .padding(.horizontal, 36)
